+++ R/03-importData-loadImport.R
@@ -1,7 +1,8 @@
 loadImport <- function(importType, params, values, filename) {
   res <- switch(importType,
          "data" = do.call(loadDataWrapper, params),
-         "model" = do.call(loadModel, params))
+         "model" = do.call(loadModel, params)) %>%
+    tryCatchWithWarningsAndErrors(errorTitle = "Could not load file!")
 
   if (importType == "model") {
     extractValuesFromModel(res, values = values, filename = filename)
@@ -32,11 +33,37 @@ selectImportParams <- function(importType,
          "model" = list(filepath = params$dataSource$file,
                         subFolder = params$subFolder,
                         rPackageName = params$rPackageName,
-                        onlySettings = params$onlySettings)
+                        onlySettings = params$onlySettings,
+                        fileExtension = params$fileExtension)
   )
 }
 
+#' Extract Values From Model
+#'
+#' Format list of model data to be compatible with the import module such that success, warnings and
+#'  errors are displayed inside the pop-up modal.
+#'
+#' @param importedModel (list) output of loadModel()
+#' @param values (reactiveValues) empty list of values in the format of the output of loadDataWrapper
+#' @param filename (reactive) name of the loaded file
+#'
+#' @return (list) list of values in the format of the output of loadDataWrapper
+extractValuesFromModel <- function(importedModel, values, filename) {
+  values$dataImport <- importedModel[c("data", "inputs", "model")]
+  values$fileName <- filename
 
+  success <- importedModel$message[importedModel$messageType == "success"]
+  warnings <- importedModel$message[importedModel$messageType == "warning"]
+  errors <- importedModel$message[importedModel$messageType == "error"]
+
+  values$fileImportSuccess <- success
+  values$warnings <- list(load = warnings)
+  values$errors <- list(load = errors)
+
+  values
+}
+
+# Functions to IMPORT DATA files ----
 
 #' Load Data Wrapper
 #'
@@ -207,6 +234,8 @@ loadData <-
     return(data)
   }
 
+# Functions to IMPORT MODEL objects ----
+
 #' Load Model
 #'
 #' @param filepath (character) path to the model file
@@ -215,42 +244,61 @@ loadModel <-
   function(filepath,
            subFolder,
            rPackageName,
-           onlySettings) {
-    dat <- list(
-      data = NULL,
-      inputs = NULL,
-      model = NULL,
-      message = c(),
-      alertType = "success",
-      uploadedVersion = ""
-    )
+           onlySettings,
+           fileExtension = "zip") {
+    expectedExt <- function(fileExtension) {
+      if (fileExtension == "zip") {
+        return(".zip")
+      } else {
+        return(sprintf(".%s or .zip", fileExtension))
+      }
+    }
 
+    # General checks: import is a valid model import ----
+    # check if valid app-specific extension or a zip file
+    if (file_ext(filepath) != fileExtension && file_ext(filepath) != "zip") {
+      stop(sprintf("Not a .%s file!", expectedExt(fileExtension)))
+      return(NULL)
+    }
+
+    ## unzip file ----
     res <- try({
-      zip::unzip(filepath)
-      modelImport <- readRDS("model.rds")
-    })
+      zip::unzip(filepath, exdir = "unzippedTmp")
+      modelImport <- extractModelFromFile(pathToUnzipped = "unzippedTmp")
+    }, silent = TRUE)
 
-    # import failures ----
+    # clean up
+    unlink("unzippedTmp", recursive = TRUE)
+
+    ## import failures
     if (inherits(res, "try-error")) {
       stop(
-        paste(
-          "The file must be a .zip file that contains the following files:",
+        sprintf(paste(
+          "The file must be a %s file that contains the following files:",
           "help.html, model.rds, README.txt.",
           "If you download a model it will have exactly this format."
-        )
+        ), expectedExt(fileExtension))
       )
       return(NULL)
     }
 
-    if (!exists("modelImport") ||
-        !all(names(modelImport) %in% c("data", "inputs", "values", "model", "version"))) {
+    if (!exists("modelImport") || !(
+      # expected names for most apps:
+      all(names(modelImport) %in% c("data", "inputs", "values", "model", "version")) ||
+      # expected names for "mpiBpred"
+      all(names(modelImport) %in% c("dataObj", "formulasObj", "inputObj", "model"))
+        )) {
       stop("File format not valid. Model object not found.")
       return(NULL)
     }
 
-    if (!is.null(rPackageName) &&
-        (rPackageName != "") &&
-        !grepl(rPackageName, modelImport$version)) {
+    # check if import was downloaded from the correct app
+
+    ## extract the name of the package without version numbers
+    rPackageLoaded <- gsub("[^a-zA-Z]", "", modelImport$version)
+
+    if (!is.null(rPackageName) && (rPackageName != "") &&
+        length(rPackageLoaded) > 0 && rPackageLoaded != "" && !grepl(rPackageName, modelImport$version)) {
       stop(
         paste(
           "Wrong model loaded. Trying to upload",
@@ -263,6 +311,8 @@ loadModel <-
       return(NULL)
     }
 
+    # Currently, this check is relevant for the iso-app where sub-models are stored in sub-folders
+    # and the name of the sub-model is kept inside $version.
     if (!is.null(rPackageName) &&
         (rPackageName != "") &&
         !is.null(subFolder) &&
@@ -283,61 +333,31 @@ loadModel <-
       return(NULL)
     }
 
-    # import checks ----
+    # Data check: which data objects are available (data, inputs and model?) ----
+    dat <- list(
+      data = NULL,
+      inputs = NULL,
+      model = NULL,
+      message = c(),
+      alertType = "success",
+      uploadedVersion = ""
+    )
+
+    dat$data <- extractDataFromModel(modelImport, rPackageName)
     ## check data ----
-    if ((rPackageName != "ReSources" && is.null(modelImport$data)) ||
-        (rPackageName == "ReSources" && is.null(modelImport$values))) {
-      dat$message[["data"]] <-
-        "No input data found."
-      dat$messageType[["data"]] <-
-        "warning"
-      dat$alertType <- "warning"
-    } else {
-      if (rPackageName == "ReSources") {
-        # loads "data" and "inputs", both stored in "values"
-        dat$data <- modelImport$values
-      } else {
-        dat$data <- modelImport$data
-      }
-      dat$message[["data"]] <-
-        "Input data loaded. "
-      dat$messageType[["data"]] <-
-        "success"
-      # no update of alertType, do not overwrite a possible warning
-    }
+    dat <- dat %>%
+      updateMessage(element = "data", msgString = "input data")
 
+    dat$inputs <- extractInputsFromModel(modelImport, rPackageName)
     ## check inputs ----
-    if ((rPackageName != "ReSources" && is.null(modelImport$data)) ||
-        (rPackageName == "ReSources" && is.null(modelImport$values))) {
-      dat$message[["inputs"]] <-
-        "No model selection parameters found."
-      dat$messageType[["inputs"]] <-
-        "warning"
-      dat$alertType <- "warning"
-    } else {
-      dat$inputs <- modelImport$inputs
-      dat$message[["inputs"]] <-
-        "Model selection parameters loaded. "
-      dat$messageType[["inputs"]] <-
-        "success"
-      # no update of alertType, do not overwrite a possible warning
-    }
+    dat <- dat %>%
+      updateMessage(element = "inputs", msgString = "model selection parameters")
 
-    ## check model ----
-    if (!onlySettings) {
-      if (is.null(modelImport$model)) {
-        dat$message[["model"]] <- "No model results found. "
-        dat$messageType[["model"]] <-
-          "warning"
-        dat$alertType <- "warning"
-      } else {
-        dat$model <- modelImport$model
-        dat$message[["model"]] <-
-          "Model results loaded. "
-        dat$messageType[["model"]] <-
-          "success"
-        # no update of alertType, do not overwrite a possible warning
-      }
+    if (!onlySettings || rPackageName %in% c("OsteoBioR")) {
+      dat$model <- extractModelFromModel(modelImport)
+      ## check model ----
+      dat <- dat %>%
+        updateMessage(element = "model", msgString = "model results")
     }
 
     if (!is.null(modelImport$version)) {
@@ -345,36 +365,105 @@ loadModel <-
         paste("Saved version:", modelImport$version, ".")
     }
 
-    # import return ----
-    # clean up
-    if (file.exists("model.rds"))
-      file.remove("model.rds")
-    if (file.exists("README.txt"))
-      file.remove("README.txt")
-    if (file.exists("help.html"))
-      file.remove("help.html")
-
     return(dat)
   }
 
-#' Extract Values From Model
+#' Extract Model From File
 #'
-#' @param importedModel (list) output of loadModel()
-#' @param values (reactiveValues) empty list of values in the format of the output of loadDataWrapper
-#' @param filename (reactive) name of the loaded file
+#' Extracts the model object from either a "model.rds" file or from a "model.Rdata" file.
+#' Recent model objects are only stored as "model.rds".
 #'
-#' @return (list) list of values in the format of the output of loadDataWrapper
-extractValuesFromModel <- function(importedModel, values, filename) {
-  values$dataImport <- importedModel[c("data", "inputs", "model")]
-  values$fileName <- filename
+#' @param pathToUnzipped (character) path to the folder were the model was unzipped
+extractModelFromFile <- function(pathToUnzipped) {
+  modelImport <- NULL
+  # load .rds file
+  if (file.exists(file.path(pathToUnzipped, "model.rds"))) {
+    modelImport <- readRDS(file.path(pathToUnzipped, "model.rds"))
+    return(modelImport)
+  }
 
-  success <- importedModel$message[importedModel$messageType == "success"]
-  warnings <- importedModel$message[importedModel$messageType == "warning"]
-  errors <- importedModel$message[importedModel$messageType == "error"]
+  # load .RData file
+  if (file.exists(file.path(pathToUnzipped, "model.Rdata"))) {
+    localEnv <- new.env()
+    load(file.path(pathToUnzipped, "model.Rdata"), envir = localEnv)
+    modelImport <- localEnv %>%
+      envToList()
+    return(modelImport)
+  }
 
-  values$fileImportSuccess <- success
-  values$warnings <- list(load = warnings)
-  values$errors <- list(load = errors)
+  modelImport
+}
 
-  values
+envToList <- function(envir) {
+  mget(x = ls(envir = envir), envir = envir)
+}
+
+extractDataFromModel <- function(modelImport, rPackageName) {
+  if (is.null(rPackageName) || rPackageName == "") return(modelImport$data)
+
+  placeholder <- list()
+  attr(placeholder, "note") <- "Find data under $model."
+
+  switch(rPackageName,
+         "ReSources" = modelImport$values,
+         "OsteoBioR" = placeholder,
+         "mpiBpred" = modelImport$dataObj,
+         "PlotR" = placeholder,
+         modelImport$data)
+}
+
+extractInputsFromModel <- function(modelImport, rPackageName) {
+  if (is.null(rPackageName) || rPackageName == "") return(modelImport$inputs)
+
+  # in ReSources model inputs are stored inside modelImport$values. Even if modelImport$inputs
+  # are empty, the inputs are loaded with dat$data <- modelImport$values.
+  placeholder <- list()
+  attr(placeholder, "note") <- switch(rPackageName,
+                                      "ReSources" = "Find input values under $data.",
+                                      "OsteoBioR" = "Find input values under $dmodel.",
+                                      "PlotR" = "Find input values under $dmodel.")
+
+  switch(rPackageName,
+         "ReSources" = placeholder,
+         "OsteoBioR" = placeholder,
+         "mpiBpred" = modelImport[c("formulasObj", "inputObj")],
+         "PlotR" = placeholder,
+         modelImport$inputs)
+}
+
+extractModelFromModel <- function(modelImport, rPackageName = NULL) {
+  if (is.null(rPackageName) || rPackageName == "") return(modelImport$model)
+
+  # Currently there are no special rules for different apps
+  return(modelImport$model)
+}
+
+#' Update Message
+#'
+#' @param dat (list) list containing data, results and warning messages of import
+#' @param element (character) list object, one of "data", "inputs", "model"
+#' @param msgString (character) description of element used in messages
+updateMessage <- function(dat, element = c("data", "inputs", "model"), msgString) {
+  element <- match.arg(element)
+
+  toUpperFirst <- function(someTxt) {
+    paste0(toupper(substring(someTxt, 1, 1)), substring(someTxt, 2))
+  }
+
+  if (is.null(dat[[element]])) {
+    dat$message[[element]] <-
+      sprintf("No %s found. ", msgString)
+    dat$messageType[[element]] <-
+      "warning"
+    dat$alertType <- "warning"
+  } else {
+    dat$message[[element]] <-
+      sprintf("%s loaded. ", msgString) %>%
+      toUpperFirst()
+    dat$messageType[[element]] <-
+      "success"
+    # no update of alertType, do not overwrite a possible warning
+  }
+
+  return(dat)
 }